Fix warning with gcc no OS/2.
git-svn-id: https://svn.netlabs.org/repos/dwindows/trunk@1677 81767d24-ef19-dc11-ae90-00e081727c95

+++ os2/dw.c
@@ -6225,7 +6225,7 @@ HWND API dw_bitmapbutton_new(char *text, ULONG id)
                    /* Check the old style header */
                    PBITMAPINFOHEADER bi = (PBITMAPINFOHEADER)&(header->bmp2);
 
-                   dw_window_set_data(tmp, "_dw_button_icon_width", DW_INT_TO_POINTER(bi->cx));
+                   dw_window_set_data(tmp, "_dw_button_icon_width", DW_INT_TO_POINTER(((int)bi->cx)));
                }
            }
            DosFreeResource(ResPtr);
